[REFACTOR] : 챌린지 진행중/완료(ChallengeHistoryDetailVC) 레이아웃 구조 변경
miniMap의 확대버튼을 통한 상세지도 정보 전환(Push NavigationController)을 위한 TableView -> baseScrollView 방식의 레이아웃 구성 요건 리팩토링
